requirements.txt
pandas==2.2.3
PyYAML==6.0.2
requests==2.32.3
python-dotenv==1.0.1
pydantic==2.10.4
pydantic-settings==2.7.0
networkx==3.4.2
rich==13.9.4
